refactor: groupe la manipulation d'objets de Publicodes dans un fichier utils dédié
