+++ site/source/pages/simulateurs/chômage-partiel/ChômagePartiel.tsx
@@ -20,7 +20,7 @@ import { H2 } from '@/design-system/typography/heading'
 import { Link } from '@/design-system/typography/link'
 import { Li, Ul } from '@/design-system/typography/list'
 import { Body } from '@/design-system/typography/paragraphs'
-import { catchDivideByZeroError } from '@/utils'
+import { catchDivideByZeroError } from '@/utils/publicodes'
 
 declare global {
 	interface Window {
